requirements.txt
python-sounddevice
numpy
scipy